Move MongoDBDependentTest class from subs-test-categories to project.
