+++ src/test/java/uk/ac/ebi/subs/apisupport/ApiSupportServiceMarkSubmittedTest.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import uk.ac.ebi.subs.ApiSupportApplication;
-import uk.ac.ebi.subs.MongoDBDependentTest;
+import uk.ac.ebi.subs.apisupport.utils.MongoDBDependentTest;
 import uk.ac.ebi.subs.data.status.ProcessingStatusEnum;
 import uk.ac.ebi.subs.data.status.SubmissionStatusEnum;
 import uk.ac.ebi.subs.repository.model.ProcessingStatus;
@@ -23,7 +23,7 @@ import uk.ac.ebi.subs.repository.repos.submittables.SampleRepository;
 
 import java.util.stream.Stream;
 
-import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
 
 @RunWith(SpringRunner.class)
