+++ src/test/java/uk/ac/ebi/subs/apisupport/utils/MongoDBDependentTest.java
@@ -1,0 +1,6 @@
+package uk.ac.ebi.subs.apisupport.utils;
+
+/**
+  * Marker interface for tests that rely on Mongo DB
+  */
+public interface MongoDBDependentTest {}
